Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/css/css-values/calc-background-position-002-expected.txt


PASS testing background-position: calc(2px + 3px) calc(4px + 5px) 
PASS testing background-position: calc(18px - 7px) calc(19px - 7px) 
PASS testing background-position: calc(4 * 5px) calc(6px * 4) 
PASS testing background-position: calc(100px / 4) calc(119px / 7) 
FAIL testing background-position: calc(6px + 21%) calc(7em + 22%) assert_equals: expected "calc(21% + 6px) calc(22% + 112px)" but got "calc(6px + 21%) calc(112px + 22%)"
FAIL testing background-position: calc(-8px + 23%) calc(-9em + 24%) assert_equals: expected "calc(23% - 8px) calc(24% - 144px)" but got "calc(-8px + 23%) calc(-144px + 24%)"
